+++ HISTORY.rst
@@ -8,6 +8,8 @@ v0.23 (unreleased)
 * improve documentation for contributing section, # 441 @pilosus
 * improve README.rst to include essential information about the package, #446 by @pilosus
 * ``IntEnum`` support, #444 by @potykion
+* fix ``ForwardRef`` collection bug, #450 by @tigerwings
+* Support specialized ``ClassVars``, #455 by @tyrylu
 * add SecretStr and SecretBytes types, #452 by @atheuz
 
 v0.22 (2019-03-29)