+++ HISTORY.rst
@@ -9,6 +9,8 @@ v0.23 (unreleased)
 * improve README.rst to include essential information about the package, #446 by @pilosus
 * ``IntEnum`` support, #444 by @potykion
 * fix PyObject callable value, #409 by @pilosus
+* fix ``ForwardRef`` collection bug, #450 by @tigerwings
+* Support specialized ``ClassVars``, #455 by @tyrylu
 
 v0.22 (2019-03-29)
 ....................